+++ lib/src/command/path.dart
@@ -1,6 +1,6 @@
 import 'package:ff_annotation_route/src/command/command.dart';
 
-class Path extends Command {
+class Path extends Command with CommandValue {
   @override
   String get description => 'The path of folder to be executed with commands.';
 
@@ -10,15 +10,12 @@ class Path extends Command {
   @override
   String get short => '-p';
 
-  /// value of path
-  String value;
-
   @override
   String toString() {
-    return '$short $value';
+    return '$full $value';
   }
 
   //help of command
   @override
-  String get command => '$short${' ' * (3 - short.length)}, $full [arguments]';
+  String get command => '$short${' ' * (3 - short.length)} , $full [arguments]';
 }